refactor(security): remove dead ValidateAPIKey, fix mustGenerateJTI, DRY path getters
- Delete ValidateAPIKey from jwt.go: zero production callers, unrelated
  to JWT lifecycle. Also remove its test and unused crypto/subtle import.
- Fix mustGenerateJTI silent failure: panic on error instead of
  returning empty string, matching the "must" naming contract.
- Extract GetAllowedBaseDirs and GetForbiddenWorkDirs from platform
  files to path.go — identical implementations were duplicated across
  path_unix.go and path_windows.go.

Fixes #341
